extra-analyse-composite-video/process_image_data/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
future==0.18.3
iso8601==2.0.0
kiwisolver==1.4.4
numpy==1.24.3
packaging==23.1
pygame==2.4.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
xmodem==0.4.6
